backend/requirements.txt
# Core dependencies
fastapi==0.110.0
uvicorn[standard]==0.27.1
python-dotenv==1.0.1

# OpenAI LLM API
openai==1.10.0

# Evaluation & Metrics
deepeval==0.20.4

# Redis for caching summaries
redis==5.0.1

# Logging
loguru==0.7.2

# NLP processing
nltk==3.8.1
spacy==3.7.2
langchain==0.1.12  # Required for embeddings
langchain_openai==0.1.0  # Supports OpenAI embeddings

# Machine Learning utilities
numpy==1.26.4
scipy==1.11.4
scikit-learn==1.3.2

# Pydantic for data validation
pydantic==2.6.1  # Upgrade to latest stable minor release

# HTTP requests
requests==2.32.3
httpx==0.27.2

# Utility libraries
tqdm==4.67.1
tenacity==8.5.0
dataclasses-json==0.6.7

# Unit Testing & Code Formatting
pytest==7.4.3 
pytest-asyncio==0.23.4
black==23.3.0
isort==5.13.2

# Production & Deployment
torch==2.2.0  # Required for embedding models
docker==7.0.0  # Required for Docker builds
flask-cors==4.0.0  # Optional: If you have CORS issues
